feat: ArticleSkeleton 내에서 분기처리
